gesture-controlled-ai-art/Python/requirements.txt
mediapipe==0.9.0
diffusers==0.21.0
opencv-python==4.5.5
torch==2.0.1
numpy==1.25.0
pillow==8.4.0
